third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gather.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] gather float32 2D tensor and int32 0D out-of-bound positive indices default options
  assert_less_than_equal: assert_array_approx_equals_ulp: test gather float32 actual -66.05901336669922 should be close enough to expected 90.2870101928711 by ULP distance: expected a number less than or equal to 0n but got 2235085098n
Harness: the test ran to completion.
